+++ pytorch/cppbuild.sh
@@ -97,12 +97,8 @@ case $PLATFORM in
         export CXX="g++ -m64"
         ;;
     macosx-*)
-        ln -sf pytorch/torch/lib ../lib
-        cp /usr/local/lib/libomp.dylib ../lib/libiomp5.dylib
-        chmod +w ../lib/libiomp5.dylib
-        install_name_tool -id @rpath/libiomp5.dylib ../lib/libiomp5.dylib
-        export CC="clang -L$INSTALL_PATH/lib -Wl,-rpath,$INSTALL_PATH/lib -liomp5 -Wno-unused-command-line-argument"
-        export CXX="clang++ -L$INSTALL_PATH/lib -Wl,-rpath,$INSTALL_PATH/lib -liomp5 -Wno-unused-command-line-argument"
+        export CC="clang"
+        export CXX="clang++"
         ;;
     windows-x86_64)
         export CC="cl.exe"
@@ -152,5 +148,15 @@ rm -Rf ../lib
 ln -sf pytorch/torch/include ../include
 ln -sf pytorch/torch/lib ../lib
 ln -sf pytorch/torch/bin ../bin
+
+# fix library with correct rpath on Mac
+case $PLATFORM in
+    macosx-*)
+        cp /usr/local/lib/libomp.dylib ../lib/libiomp5.dylib
+        chmod +w ../lib/libiomp5.dylib
+        install_name_tool -id @rpath/libiomp5.dylib ../lib/libiomp5.dylib
+        install_name_tool -change @rpath/libomp.dylib @rpath/libiomp5.dylib ../lib/libtorch_cpu.dylib
+        ;;
+esac
 
 cd ../..